+++ Altinn/AT.Common.Altinn.Test/Unit/Setup/AltinnApiTestFixture.cs
@@ -3,7 +3,7 @@ using System.Text.Json;
 using Arbeidstilsynet.Common.Altinn.DependencyInjection;
 using Arbeidstilsynet.Common.Altinn.Model.Api;
 using Arbeidstilsynet.Common.Altinn.Ports.Token;
-using Arbeidstilsynet.Common.TestExtensions.Extensions;
+using Arbeidstilsynet.Shared.TestExtensions;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;