+++ Altinn/AT.Common.Altinn.Publish/CHANGELOG.md
@@ -19,6 +19,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security <!-- in case of vulnerabilities. -->
 
+## 3.0.0
+
+### Removed
+
+- **BREAKING**: Removed dependency on `Arbeidstilsynet.Common.AspNetCore.Extensions`
+
+### Changed
+
+- Added `FrameworkReference` to `Microsoft.AspNetCore.App` for ASP.NET Core types
+
 ## 2.6.1
 
 ### Changed